[SCAL-135040] Move Diamond from Python 2 to Python 3 syntax

+++ src/collectors/proc/proc.py
@@ -64,7 +64,7 @@ class ProcessStatCollector(diamond.collector.Collector):
                 metric_name = data[0]
                 metric_value = int(data[1])
                 metric_value = int(self.derivative(metric_name,
-                                                   long(metric_value),
+                                                   int(metric_value),
                                                    counter))
                 self.publish(metric_name, metric_value)
